backend/pyproject.toml
[project]
name = "videometa-backend"
version = "0.1.0"
description = "影片自動化標籤與摘要系統 - 後端服務"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "python-multipart>=0.0.9",
    "google-genai>=1.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.0.0",
    "openpyxl>=3.1.0",
    "aiofiles>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
